Refactor DatabaseService into DatabaseManager
